+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 -------------------
 
 - When an annex has been modified, avoid to reindex the entire parent, only reindex relevant indexes : modified related indexes and SearchableText as annex title is indexed into it
+- Integrated new column "publishable" from collective.iconifiedcategory
 
 4.1.9 (2019-11-04)
 ------------------